Conflicts:
	Banshee.sln

+++ src/Dap/Banshee.Dap.MassStorage/Banshee.Dap.MassStorage/MassStorageSource.cs
@@ -38,6 +38,7 @@ using Banshee.Dap;
 using Banshee.Base;
 using Banshee.ServiceStack;
 using Banshee.Library;
+using Banshee.Query;
 using Banshee.Sources;
 using Banshee.Collection;
 using Banshee.Collection.Database;
@@ -164,18 +165,18 @@ namespace Banshee.Dap.MassStorage
             if (CanSyncPlaylists) {
                 var insert_cmd = new Hyena.Data.Sqlite.HyenaSqliteCommand (
                     "INSERT INTO CorePlaylistEntries (PlaylistID, TrackID) VALUES (?, ?)");
-                int [] psources = new int [] {DbId};
                 foreach (string playlist_path in PlaylistFiles) {
                     IPlaylistFormat loaded_playlist = PlaylistFileUtil.Load (playlist_path, new Uri (PlaylistsPath));
                     if (loaded_playlist == null)
                         continue;
 
-                    PlaylistSource playlist = new PlaylistSource (System.IO.Path.GetFileNameWithoutExtension (playlist_path), this);
+                    string name = System.IO.Path.GetFileNameWithoutExtension (SafeUri.UriToFilename (playlist_path));
+                    PlaylistSource playlist = new PlaylistSource (name, this);
                     playlist.Save ();
                     //Hyena.Data.Sqlite.HyenaSqliteCommand.LogAll = true;
                     foreach (Dictionary<string, object> element in loaded_playlist.Elements) {
                         string track_path = (element["uri"] as Uri).LocalPath;
-                        int track_id = DatabaseTrackInfo.GetTrackIdForUri (new SafeUri (track_path), psources);
+                        int track_id = DatabaseTrackInfo.GetTrackIdForUri (new SafeUri (track_path), DbId);
                         if (track_id == 0) {
                             Log.DebugFormat ("Failed to find track {0} in DAP library to load it into playlist {1}", track_path, playlist_path);
                         } else {
@@ -328,7 +329,8 @@ namespace Banshee.Dap.MassStorage
                 if (from != null && !String.IsNullOrEmpty (escaped_name)) {
                     from.Reload ();
                     if (playlist_format == null) {
-                         playlist_format = Activator.CreateInstance (PlaylistTypes[0].Type) as PlaylistFormatBase;
+                        playlist_format = Activator.CreateInstance (PlaylistTypes[0].Type) as PlaylistFormatBase;
+                        playlist_format.FolderSeparator = MediaCapabilities.FolderSeparator;
                     }
 
                     SafeUri playlist_path = new SafeUri (System.IO.Path.Combine (
@@ -484,6 +486,24 @@ namespace Banshee.Dap.MassStorage
             set { cover_art_file_type = value; }
         }
 
+        public override void UpdateMetadata (DatabaseTrackInfo track)
+        {
+            SafeUri new_uri = new SafeUri (GetTrackPath (track, System.IO.Path.GetExtension (track.Uri)));
+
+            if (new_uri.ToString () != track.Uri.ToString ()) {
+                Directory.Create (System.IO.Path.GetDirectoryName (new_uri.LocalPath));
+                Banshee.IO.File.Move (track.Uri, new_uri);
+
+                //to remove the folder if it's not needed anymore:
+                DeleteTrackFile (track);
+
+                track.Uri = new_uri;
+                track.Save (true, BansheeQuery.UriField);
+            }
+
+            base.UpdateMetadata (track);
+        }
+
         protected override void AddTrackToDevice (DatabaseTrackInfo track, SafeUri fromUri)
         {
             if (track.PrimarySourceId == DbId)
@@ -514,8 +534,9 @@ namespace Banshee.Dap.MassStorage
                 if (track.LastSyncedStamp >= Hyena.DateTimeUtil.ToDateTime (track.FileModifiedStamp)) {
                     Log.DebugFormat ("Copying Metadata to File Since Sync time >= Updated Time");
                     bool write_metadata = Metadata.SaveTrackMetadataService.WriteMetadataEnabled.Value;
-                    bool write_ratings_and_playcounts = Metadata.SaveTrackMetadataService.WriteRatingsAndPlayCountsEnabled.Value;
-                    Banshee.Streaming.StreamTagger.SaveToFile (copied_track, write_metadata, write_ratings_and_playcounts);
+                    bool write_ratings = Metadata.SaveTrackMetadataService.WriteRatingsEnabled.Value;
+                    bool write_playcounts = Metadata.SaveTrackMetadataService.WritePlayCountsEnabled.Value;
+                    Banshee.Streaming.StreamTagger.SaveToFile (copied_track, write_metadata, write_ratings, write_playcounts);
                 }
 
                 copied_track.Save (false);
@@ -559,11 +580,16 @@ namespace Banshee.Dap.MassStorage
 
         protected override bool DeleteTrack (DatabaseTrackInfo track)
         {
-            try {
-                if (ms_device != null && !ms_device.DeleteTrackHook (track)) {
-                    return false;
-                }
+            if (ms_device != null && !ms_device.DeleteTrackHook (track)) {
+                return false;
+            }
+            DeleteTrackFile (track);
+            return true;
+        }
 
+        private void DeleteTrackFile (DatabaseTrackInfo track)
+        {
+            try {
                 string track_file = System.IO.Path.GetFileName (track.Uri.LocalPath);
                 string track_dir = System.IO.Path.GetDirectoryName (track.Uri.LocalPath);
                 int files = 0;
@@ -583,12 +609,14 @@ namespace Banshee.Dap.MassStorage
                     System.IO.File.Delete (Paths.Combine (track_dir, CoverArtFileName));
                 }
 
-                Banshee.IO.Utilities.DeleteFileTrimmingParentDirectories (track.Uri);
+                if (Banshee.IO.File.Exists (track.Uri)) {
+                    Banshee.IO.Utilities.DeleteFileTrimmingParentDirectories (track.Uri);
+                } else {
+                    Banshee.IO.Utilities.TrimEmptyDirectories (track.Uri);
+                }
             } catch (System.IO.FileNotFoundException) {
             } catch (System.IO.DirectoryNotFoundException) {
             }
-
-            return true;
         }
 
         protected override void Eject ()
@@ -631,6 +659,9 @@ namespace Banshee.Dap.MassStorage
                 // the exact given depth (not including the mount point/audio_folder).
                 if (FolderDepth != -1) {
                     int depth = FolderDepth;
+
+                    bool is_album_unknown = String.IsNullOrEmpty (track.AlbumTitle);
+
                     string album_artist = FileNamePattern.Escape (track.DisplayAlbumArtistName);
                     string track_album  = FileNamePattern.Escape (track.DisplayAlbumTitle);
                     string track_number = FileNamePattern.Escape (String.Format ("{0:00}", track.TrackNumber));
@@ -639,15 +670,21 @@ namespace Banshee.Dap.MassStorage
                     if (depth == 0) {
                         // Artist - Album - 01 - Title
                         string track_artist = FileNamePattern.Escape (track.DisplayArtistName);
-                        file_path = String.Format ("{0} - {1} - {2} - {3}", track_artist, track_album, track_number, track_title);
+                        file_path = is_album_unknown ?
+                            String.Format ("{0} - {1} - {2}", track_artist, track_number, track_title) :
+                            String.Format ("{0} - {1} - {2} - {3}", track_artist, track_album, track_number, track_title);
                     } else if (depth == 1) {
                         // Artist - Album/01 - Title
-                        file_path = String.Format ("{0} - {1}", album_artist, track_album);
+                        file_path = is_album_unknown ?
+                            album_artist :
+                            String.Format ("{0} - {1}", album_artist, track_album);
                         file_path = System.IO.Path.Combine (file_path, String.Format ("{0} - {1}", track_number, track_title));
                     } else if (depth == 2) {
                         // Artist/Album/01 - Title
                         file_path = album_artist;
-                        file_path = System.IO.Path.Combine (file_path, track_album);
+                        if (!is_album_unknown || ms_device.MinimumFolderDepth == depth) {
+                            file_path = System.IO.Path.Combine (file_path, track_album);
+                        }
                         file_path = System.IO.Path.Combine (file_path, String.Format ("{0} - {1}", track_number, track_title));
                     } else {
                         // If the *required* depth is more than 2..go nuts!
@@ -678,6 +715,10 @@ namespace Banshee.Dap.MassStorage
             file_path += ext;
 
             return file_path;
+        }
+
+        public override bool HasEditableTrackProperties {
+            get { return true; }
         }
     }
 }
